Revert t variable type to float instead of double For the sake of simplicity and consistency.

+++ flix/mavlink.ino
@@ -15,7 +15,7 @@
 bool mavlinkConnected = false;
 String mavlinkPrintBuffer;
 
-extern double controlTime;
+extern float controlTime;
 extern float controlRoll, controlPitch, controlThrottle, controlYaw, controlMode;
 
 void processMavlink() {
@@ -26,8 +26,8 @@ void processMavlink() {
 void sendMavlink() {
 	sendMavlinkPrint();
 
-	static double lastSlow = 0;
-	static double lastFast = 0;
+	static float lastSlow = 0;
+	static float lastFast = 0;
 
 	mavlink_message_t msg;
 	uint32_t time = t * 1000;
